Fixed warnings going from size_t to int
git-svn-id: https://unittest-cpp.svn.sourceforge.net/svnroot/unittest-cpp@115 feadf2a4-a60d-0410-a25d-8df00a7d1702

+++ UnitTest++/src/MemoryOutStream.cpp
@@ -42,8 +42,8 @@ char const* MemoryOutStream::GetText() const
 
 MemoryOutStream& MemoryOutStream::operator << (char const* txt)
 {
-    int const bytesLeft = m_capacity - std::strlen(m_buffer);
-    int const bytesRequired = std::strlen(txt) + 1;
+    int const bytesLeft = m_capacity - (int)std::strlen(m_buffer);
+    int const bytesRequired = (int)std::strlen(txt) + 1;
 
     if (bytesRequired > bytesLeft)
     {
